refactor(app): extract palette hooks into usePaletteWiring
- Move the ~10 palette hooks (new-terminal, panel, project-switcher,
  action, quick-switcher, send-to-agent, worktree, quick-create) plus
  useDoubleShift/useProjectMruSwitcher, the theme/log-level/
  resume-sessions palette-store flags, and every palette's
  useKeepMounted exit-animation latch out of AppInner into one hook
- Non-palette useKeepMounted latches (cross-diff dialog, shortcuts
  dialog) stay inline in AppInner — they aren't palettes
- Reorders these unconditional hook calls relative to a few unrelated
  dialog-state hooks (settings, plugin manager, git-init); no
  conditional hooks and no cross-hook data dependency, so call order
  stays legal and behavior is unchanged
- No logic changes
